+++ lib/qsu.dart
@@ -1,5 +1,6 @@
 library;
 
 export 'src/format.dart';
+export 'src/misc.dart';
 export 'src/string.dart';
 export 'src/verify.dart';