+++ lib/src/misc.dart
@@ -1,0 +1,3 @@
+Future<void> sleep(int delay) async {
+  await Future.delayed(Duration(milliseconds: delay));
+}