+++ test/misc_test.dart
@@ -1,0 +1,10 @@
+import 'package:qsu/qsu.dart';
+import 'package:test/test.dart';
+
+void main() {
+  group('Misc', () {
+    test('sleep', () {
+      expect(sleep(100), completes);
+    });
+  });
+}